+++ src/main/java/com/github/Cubolink/finalreality/model/weapon/Axe.java
@@ -1,7 +1,7 @@
 package com.github.Cubolink.finalreality.model.weapon;
 
 public class Axe extends AbstractWeapon{
-    public Axe(String name, double physical_damage, double weight) {
+    public Axe(String name, int physical_damage, double weight) {
         super(name, physical_damage, weight);
     }
 }
