+++ src/main/java/com/github/Cubolink/finalreality/model/weapon/Bow.java
@@ -1,7 +1,7 @@
 package com.github.Cubolink.finalreality.model.weapon;
 
 public class Bow extends AbstractWeapon{
-    public Bow(String name, double physical_damage, double weight) {
+    public Bow(String name, int physical_damage, double weight) {
         super(name, physical_damage, weight);
     }
 }
