+++ src/main/java/com/github/Cubolink/finalreality/model/weapon/Staff.java
@@ -1,15 +1,15 @@
 package com.github.Cubolink.finalreality.model.weapon;
 
 public class Staff extends AbstractWeapon{
-    protected final double magical_damage;
+    protected final int magical_damage;
 
-    public Staff(String name, double physical_damage, double weight, double magical_damage) {
+    public Staff(String name, int physical_damage, double weight, int magical_damage) {
         super(name, physical_damage, weight);
         this.magical_damage = magical_damage;
     }
 
     @Override
-    public double getMagical_damage(){
+    public int getMagical_damage(){
         return magical_damage;
     }
 }
