+++ src/main/java/com/github/Cubolink/finalreality/model/weapon/Sword.java
@@ -1,7 +1,7 @@
 package com.github.Cubolink.finalreality.model.weapon;
 
 public class Sword extends AbstractWeapon{
-    public Sword(String name, double physical_damage, double weight) {
+    public Sword(String name, int physical_damage, double weight) {
         super(name, physical_damage, weight);
     }
 }
